+++ src/tools/ta_regtest/ta_test_func/test_avgdev.c
@@ -256,7 +256,7 @@ static ErrorNumber do_test( const TA_History *history,
    {
       errNb = doRangeTest(
                            rangeTestFunction,
-                           TA_FUNC_UNST_NONE,
+                           TA_TEST_UNST_NONE,
                            (void *)&testParam, 1, 0 );
       if( errNb != TA_TEST_PASS )
          return errNb;